enhance thread ui
